Auto poll maxInitWaitTimeSeconds fix
